Split up forms.tsx into multiple files. (#709)
