+++ frontend/lib/hp-action.tsx
@@ -6,7 +6,7 @@ import { CenteredPrimaryButtonLink, NextButton, ProgressButtons } from './button
 import { IssuesRoutes } from './pages/issue-pages';
 import { withAppContext, AppContextType } from './app-context';
 import { AllSessionInfo_landlordDetails, AllSessionInfo, AllSessionInfo_feeWaiver } from './queries/AllSessionInfo';
-import { SessionUpdatingFormSubmitter, FormContextRenderer } from './forms';
+import { SessionUpdatingFormSubmitter } from './session-updating-form-submitter';
 import { GenerateHPActionPDFMutation } from './queries/GenerateHPActionPDFMutation';
 import { PdfLink } from './pdf-link';
 import { ProgressRoutesProps, buildProgressRoutesComponent } from './progress-routes';
@@ -27,6 +27,7 @@ import { YesNoRadiosFormField } from './yes-no-radios-form-field';
 import { SessionStepBuilder } from './session-step-builder';
 import { HpActionSueForHarassmentMutation } from './queries/HpActionSueForHarassmentMutation';
 import { HarassmentApartment, HarassmentExplain, HarassmentCaseHistory, HarassmentAllegations1, HarassmentAllegations2 } from './pages/hp-action-harassment';
+import { FormContextRenderer } from './form';
 
 const onboardingForHPActionRoute = () => Routes.locale.hp.onboarding.latestStep;
