+++ frontend/lib/session-step-builder.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { SessionUpdatingFormOutput, FormContext, SessionUpdatingFormSubmitter } from "./forms";
+import { SessionUpdatingFormOutput, SessionUpdatingFormSubmitter } from "./session-updating-form-submitter";
 import { FetchMutationInfo } from "./forms-graphql";
 import { FormInputConverter, getInitialFormInput } from "./form-input-converter";
 import { MiddleProgressStep, MiddleProgressStepProps } from "./progress-step-route";
@@ -8,6 +8,7 @@ import Page from "./page";
 import { ProgressButtons } from './buttons';
 import { AllSessionInfo } from './queries/AllSessionInfo';
 import { WithServerFormFieldErrors } from './form-errors';
+import { FormContext } from './form-context';
 
 /**
  * A structure, usually created by querybuilder, which contains information
